+++ .gitignore
@@ -44,3 +44,6 @@ src/main/resources/application-*.yml
 
 # QueryDSL이 생성하는 Q클래스 경로
 src/main/generated/
+
+# Windows 예약어 파일 (자동 생성 방지)
+nul
