+++ src/main/java/kr/eolmago/global/config/SecurityConfig.java
@@ -41,7 +41,6 @@ public class SecurityConfig {
                         "/swagger-ui.html",
                         "/v3/api-docs/**",
                         "/api-docs/**",
-
                         "/api/**"
                 )
                 .csrf(AbstractHttpConfigurer::disable)
